refactor(utils): fold card_images.py into image_service mixins (#442)
Splits the 1104-line utils/card_images.py across new modules under
services/image_service/: schemas (constants/decoders/CardImageRequest),
path_resolver, disk_cache (CardImageCache), downloader (BulkImageDownloader
+ high-level helpers). ImageService keeps the same mixin composition.

All 11 external import sites switched from utils.card_images to
services.image_service; public API names preserved. utils/card_images.py
deleted.
