+++ services/image_service/cache.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 
 from loguru import logger
 
-from utils.card_images import CardImageRequest
+from services.image_service.schemas import CardImageRequest
 
 if TYPE_CHECKING:
     from services.image_service.protocol import ImageServiceProto
