--- utils/card_images.py
@@ -1,1104 +1,0 @@
-"""High-throughput card image downloader and cache manager.
-
-This module provides functionality to:
-1. Download Scryfall bulk data to get all card metadata
-2. Fetch card images from Scryfall CDN (no rate limits on image CDN)
-3. Maintain a local SQLite database of cached images
-4. Support concurrent downloads for high throughput
-
-Architecture:
-- Uses Scryfall bulk data JSON for card metadata
-- Downloads images from cards.scryfall.io CDN (no rate limits)
-- Stores images locally with UUID-based filenames
-- SQLite database tracks downloaded images and metadata
-- Supports multiple image sizes (small, normal, large, png)
-"""
-
-from __future__ import annotations
-
-import os
-import sqlite3
-import threading
-import unicodedata
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass
-from datetime import datetime, timezone
-
-import msgspec
-import msgspec.json
-
-try:  # Python 3.11+ has UTC
-    from datetime import UTC
-except ImportError:  # pragma: no cover - compatibility shim for Python 3.10
-    UTC = timezone.utc  # noqa: UP017
-from pathlib import Path, PureWindowsPath
-from typing import Any
-
-import requests
-from loguru import logger
-
-from utils.atomic_io import (
-    atomic_write_bytes,
-    atomic_write_json,
-    atomic_write_stream,
-    locked_path,
-)
-from utils.constants import (
-    BULK_DATA_CACHE_FRESHNESS_SECONDS,
-    BYTES_PER_MB,
-    CACHE_DIR,
-    SCRYFALL_BULK_STREAM_TIMEOUT_SECONDS,
-    SCRYFALL_DOWNLOAD_CHUNK_SIZE,
-    SCRYFALL_DOWNLOAD_PROGRESS_INTERVAL,
-    SCRYFALL_MAX_DOWNLOAD_WORKERS,
-    SCRYFALL_REQUEST_TIMEOUT_SECONDS,
-    SQLITE_CONNECTION_TIMEOUT_SECONDS,
-)
-from utils.perf import timed
-
-# Image cache configuration
-IMAGE_CACHE_DIR = CACHE_DIR / "card_images"
-IMAGE_DB_PATH = IMAGE_CACHE_DIR / "images.db"
-BULK_DATA_CACHE = IMAGE_CACHE_DIR / "bulk_data.json"
-PRINTING_INDEX_VERSION = 3
-PRINTING_INDEX_CACHE = IMAGE_CACHE_DIR / f"printings_v{PRINTING_INDEX_VERSION}.json"
-
-# Image size options (in order of preference for storage)
-IMAGE_SIZES = {
-    "small": "small",  # 146x204 - thumbnails
-    "normal": "normal",  # 488x680 - default
-    "large": "large",  # 672x936 - high quality
-    "png": "png",  # 745x1040 - highest quality, transparent
-}
-
-# Download configuration
-BULK_DATA_URL = "https://api.scryfall.com/bulk-data/default-cards"
-SCRYFALL_CARD_NAMED_URL = "https://api.scryfall.com/cards/named"
-SCRYFALL_CARD_SEARCH_URL = "https://api.scryfall.com/cards/search"
-
-
-# ---------------------------------------------------------------------------
-# msgspec schemas for fast JSON loading
-# ---------------------------------------------------------------------------
-
-
-class BulkCardFace(msgspec.Struct, gc=False):
-    """Minimal face entry from Scryfall bulk data (only fields we use)."""
-
-    name: str | None = None
-    image_uris: dict[str, str] | None = None
-
-    def get(self, key: str, default: Any = None) -> Any:  # noqa: ANN401
-        return getattr(self, key, default)
-
-    def __getitem__(self, key: str) -> Any:  # noqa: ANN401
-        try:
-            return getattr(self, key)
-        except AttributeError:
-            raise KeyError(key) from None
-
-
-class BulkCard(msgspec.Struct, gc=False):
-    """Minimal Scryfall bulk-data card record containing only the fields
-    required by :func:`build_printing_index`.  msgspec silently skips all
-    other fields present in the JSON, which makes parsing even faster.
-
-    Dict-compatible accessors are provided so that the existing
-    ``card.get(...)`` call sites continue to work unchanged.
-    """
-
-    name: str | None = None
-    id: str | None = None
-    set: str | None = None
-    set_name: str | None = None
-    collector_number: str | None = None
-    released_at: str | None = None
-    flavor_text: str | None = None
-    artist: str | None = None
-    card_faces: list[BulkCardFace] | None = None
-
-    def get(self, key: str, default: Any = None) -> Any:  # noqa: ANN401
-        return getattr(self, key, default)
-
-    def __getitem__(self, key: str) -> Any:  # noqa: ANN401
-        try:
-            return getattr(self, key)
-        except AttributeError:
-            raise KeyError(key) from None
-
-
-class PrintingEntry(msgspec.Struct, gc=False):
-    """A single card printing record stored in the printings index."""
-
-    id: str
-    set: str
-    set_name: str
-    collector_number: str
-    released_at: str
-    flavor_text: str = ""
-    artist: str = ""
-
-    def get(self, key: str, default: Any = None) -> Any:  # noqa: ANN401
-        return getattr(self, key, default)
-
-    def __getitem__(self, key: str) -> Any:  # noqa: ANN401
-        try:
-            return getattr(self, key)
-        except AttributeError:
-            raise KeyError(key) from None
-
-
-class PrintingIndexPayload(msgspec.Struct):
-    """Root structure of the saved printing index cache file."""
-
-    version: int
-    generated_at: str
-    bulk_mtime: float
-    unique_names: int
-    total_printings: int
-    data: dict[str, list[PrintingEntry]]
-
-
-# Pre-instantiated decoders – reusing avoids re-building on every call.
-_bulk_cards_decoder: msgspec.json.Decoder[list[BulkCard]] = msgspec.json.Decoder(list[BulkCard])
-_printing_index_decoder: msgspec.json.Decoder[PrintingIndexPayload] = msgspec.json.Decoder(
-    PrintingIndexPayload
-)
-
-
-@dataclass(frozen=True)
-class CardImageRequest:
-    """Represents a single card image download request."""
-
-    card_name: str
-    uuid: str | None
-    set_code: str | None
-    collector_number: str | None
-    size: str = "normal"
-
-    def queue_key(self) -> tuple[str, str, str, str]:
-        if self.uuid:
-            return ("uuid", self.uuid.lower(), self.size, "")
-        set_code = (self.set_code or "").lower()
-        collector = (self.collector_number or "").lower()
-        return ("set", set_code, collector, self.size)
-
-    def can_fetch(self) -> bool:
-        return bool((self.card_name or "").strip())
-
-
-def _strip_accents(text: str) -> str:
-    """Return *text* with combining diacritical marks removed (e.g. ó → o)."""
-    return "".join(
-        c for c in unicodedata.normalize("NFKD", text) if not unicodedata.combining(c)
-    ).lower()
-
-
-class CardImageCache:
-    """Manages local card image cache with SQLite database."""
-
-    def __init__(self, cache_dir: Path = IMAGE_CACHE_DIR, db_path: Path = IMAGE_DB_PATH):
-        self.cache_dir = Path(cache_dir)
-        self.db_path = Path(db_path)
-        self._ensure_directories()
-        self.cache_dir = self.cache_dir.resolve()
-        self.db_path = self.db_path.resolve()
-        self._path_roots = self._build_path_roots()
-        self._init_database()
-        self._path_cache: dict[tuple[str, str], Path | None] = {}
-        self._path_cache_lock: threading.Lock = threading.Lock()
-
-    def _ensure_directories(self) -> None:
-        self.cache_dir.mkdir(parents=True, exist_ok=True)
-        for size in IMAGE_SIZES.values():
-            (self.cache_dir / size).mkdir(exist_ok=True)
-
-    def _build_path_roots(self) -> list[Path]:
-        roots: list[Path] = []
-        candidates = [
-            Path.cwd(),
-            self.cache_dir,
-            self.cache_dir.parent,
-        ]
-        # Include grandparent (project root) if available
-        try:
-            candidates.append(self.cache_dir.parents[1])
-        except IndexError:
-            pass
-        seen = set()
-        for entry in candidates:
-            resolved = entry.resolve()
-            if resolved not in seen:
-                seen.add(resolved)
-                roots.append(resolved)
-        return roots
-
-    def _init_database(self) -> None:
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            self._create_schema(conn)
-            self._ensure_face_index_support(conn)
-            conn.commit()
-
-    def _create_schema(self, conn: sqlite3.Connection) -> None:
-        conn.execute("""
-            CREATE TABLE IF NOT EXISTS card_images (
-                uuid TEXT NOT NULL,
-                face_index INTEGER NOT NULL DEFAULT 0,
-                name TEXT NOT NULL,
-                set_code TEXT,
-                collector_number TEXT,
-                image_size TEXT NOT NULL,
-                file_path TEXT NOT NULL,
-                downloaded_at TEXT NOT NULL,
-                scryfall_uri TEXT,
-                artist TEXT,
-                PRIMARY KEY (uuid, face_index, image_size)
-            )
-        """)
-        conn.execute("""
-            CREATE INDEX IF NOT EXISTS idx_card_name ON card_images(name)
-        """)
-        conn.execute("""
-            CREATE INDEX IF NOT EXISTS idx_set_code ON card_images(set_code)
-        """)
-        conn.execute("""
-            CREATE TABLE IF NOT EXISTS bulk_data_meta (
-                id INTEGER PRIMARY KEY CHECK (id = 1),
-                downloaded_at TEXT NOT NULL,
-                total_cards INTEGER NOT NULL,
-                bulk_data_uri TEXT NOT NULL
-            )
-        """)
-
-    def _ensure_face_index_support(self, conn: sqlite3.Connection) -> None:
-        info = conn.execute("PRAGMA table_info(card_images)").fetchall()
-        has_face_index = any(column[1] == "face_index" for column in info)
-        if has_face_index:
-            return
-
-        logger.info("Migrating card_images table to support multi-face entries")
-        conn.execute("ALTER TABLE card_images RENAME TO card_images_old")
-        self._create_schema(conn)
-        conn.execute("""
-            INSERT INTO card_images (
-                uuid,
-                face_index,
-                name,
-                set_code,
-                collector_number,
-                image_size,
-                file_path,
-                downloaded_at,
-                scryfall_uri,
-                artist
-            )
-            SELECT
-                uuid,
-                0,
-                name,
-                set_code,
-                collector_number,
-                image_size,
-                file_path,
-                downloaded_at,
-                scryfall_uri,
-                artist
-            FROM card_images_old
-        """)
-        conn.execute("DROP TABLE card_images_old")
-
-    def _resolve_path(self, stored_path: str) -> Path:
-        """Convert stored path strings into usable filesystem Paths.
-
-        Handles Windows-style separators and WSL drive prefixes when running on POSIX.
-        """
-        raw = stored_path.strip()
-        path = Path(raw)
-        resolved = self._normalize_path(path)
-        if resolved.exists():
-            return resolved
-
-        # Normalize backslashes to forward slashes (works on all OSes)
-        if "\\" in raw:
-            normalized = raw.replace("\\", "/")
-            path = Path(normalized)
-            normalized_resolved = self._normalize_path(path)
-            if normalized_resolved.exists():
-                return normalized_resolved
-
-            # Interpret as Windows path and convert to current platform
-            try:
-                win_path = Path(PureWindowsPath(raw))
-                if win_path.exists():
-                    return win_path
-            except Exception as exc:
-                logger.debug("Failed to normalize Windows path '%s': %s", raw, exc)
-
-            # Translate Windows drive letters for WSL paths (e.g., C:\ -> /mnt/c/)
-            if os.name != "nt" and len(raw) >= 3 and raw[1] == ":" and raw[2] in ("\\", "/"):
-                drive = raw[0].lower()
-                remainder = raw[3:].replace("\\", "/")
-                wsl_path = Path("/mnt") / drive / remainder
-                if wsl_path.exists():
-                    return wsl_path
-
-        # Project may have been renamed (e.g. magic_online_metagame_crawler →
-        # mtgo_tools).  The stored absolute path no longer exists but the file
-        # itself (identified by its UUID filename) lives under the current
-        # cache_dir.  Reconstruct the path from just the last two components
-        # (size-subfolder/uuid.jpg) relative to the current cache_dir.
-        try:
-            raw_path = Path(raw.replace("\\", "/"))
-            rebased = self.cache_dir / raw_path.parts[-2] / raw_path.name
-            if rebased.exists():
-                return rebased
-        except Exception:
-            pass
-
-        return resolved
-
-    def _normalize_path(self, path: Path) -> Path:
-        try:
-            if path.is_absolute():
-                return path
-        except OSError:
-            pass
-
-        rel = self._resolve_relative_path(path)
-        if rel is not None:
-            return rel
-
-        return path
-
-    def _resolve_relative_path(self, relative: Path) -> Path | None:
-        for root in self._path_roots:
-            candidate = (root / relative).resolve()
-            if candidate.exists():
-                return candidate
-        return None
-
-    @timed
-    def get_image_path(self, card_name: str, size: str = "normal") -> Path | None:
-        key = (card_name.lower(), size)
-        with self._path_cache_lock:
-            if key in self._path_cache:
-                return self._path_cache[key]
-        result = self._get_image_path_from_db(card_name, size)
-        with self._path_cache_lock:
-            # Only cache positive hits; None means "not yet downloaded"
-            # and would suppress future availability after a download.
-            if result is not None:
-                self._path_cache[key] = result
-        return result
-
-    def _get_image_path_from_db(self, card_name: str, size: str) -> Path | None:
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            cursor = conn.execute(
-                """
-                SELECT file_path
-                FROM card_images
-                WHERE LOWER(name) = LOWER(?) AND image_size = ?
-                ORDER BY face_index
-                LIMIT 1
-                """,
-                (card_name, size),
-            )
-            row = cursor.fetchone()
-            if row:
-                path = self._resolve_path(row[0])
-                if path.exists():
-                    return path
-
-            alias_path = self._lookup_double_faced_alias(conn, card_name, size)
-            if alias_path:
-                return alias_path
-
-            # Fallback: accent-insensitive lookup for cards like "Lórien Revealed"
-            # stored under their accented Scryfall name but requested without accents
-            # (or vice-versa).  SQLite's LOWER() does not strip combining characters,
-            # so we register a custom scalar and compare normalised forms.
-            conn.create_function("strip_accents", 1, _strip_accents)
-            cursor = conn.execute(
-                """
-                SELECT file_path
-                FROM card_images
-                WHERE strip_accents(name) = ? AND image_size = ?
-                ORDER BY face_index
-                LIMIT 1
-                """,
-                (_strip_accents(card_name), size),
-            )
-            row = cursor.fetchone()
-            if row:
-                path = self._resolve_path(row[0])
-                if path.exists():
-                    return path
-        return None
-
-    def _lookup_double_faced_alias(
-        self, conn: sqlite3.Connection, card_name: str, size: str
-    ) -> Path | None:
-        alias = (card_name or "").strip()
-        if not alias or "//" in alias:
-            return None
-
-        alias_lower = alias.lower()
-        patterns = (
-            f"{alias_lower} // %",
-            f"% // {alias_lower}",
-        )
-
-        for pattern in patterns:
-            cursor = conn.execute(
-                """
-                SELECT file_path
-                FROM card_images
-                WHERE LOWER(name) LIKE ? AND image_size = ?
-                ORDER BY face_index
-                LIMIT 1
-                """,
-                (pattern, size),
-            )
-            row = cursor.fetchone()
-            if row:
-                path = self._resolve_path(row[0])
-                if path.exists():
-                    return path
-        return None
-
-    def get_image_path_for_printing(
-        self, card_name: str, set_code: str, size: str = "normal"
-    ) -> Path | None:
-        if not set_code:
-            return self.get_image_path(card_name, size)
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            cursor = conn.execute(
-                """
-                SELECT file_path
-                FROM card_images
-                WHERE LOWER(name) = LOWER(?) AND LOWER(set_code) = LOWER(?) AND image_size = ?
-                ORDER BY face_index
-                LIMIT 1
-                """,
-                (card_name, set_code, size),
-            )
-            row = cursor.fetchone()
-            if not row:
-                return None
-            return self._resolve_path(row[0])
-
-    def get_image_by_uuid(
-        self, uuid: str, size: str = "normal", face_index: int | None = 0
-    ) -> Path | None:
-        query = "SELECT file_path FROM card_images WHERE uuid = ? AND image_size = ? ORDER BY face_index"
-        params: tuple[object, ...]
-        if face_index is None:
-            params = (uuid, size)
-        else:
-            query = "SELECT file_path FROM card_images WHERE uuid = ? AND face_index = ? AND image_size = ?"
-            params = (uuid, face_index, size)
-
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            cursor = conn.execute(query, params)
-            row = cursor.fetchone()
-            if row:
-                path = self._resolve_path(row[0])
-                if path.exists():
-                    return path
-        return None
-
-    def get_image_paths_by_uuid(self, uuid: str, size: str = "normal") -> list[Path]:
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            rows = conn.execute(
-                """
-                SELECT face_index, file_path
-                FROM card_images
-                WHERE uuid = ? AND image_size = ? AND face_index >= 0
-                ORDER BY face_index
-                """,
-                (uuid, size),
-            ).fetchall()
-        paths: list[Path] = []
-        for _, file_path in rows:
-            path = self._resolve_path(file_path)
-            if path.exists():
-                paths.append(path)
-        return paths
-
-    def add_image(
-        self,
-        uuid: str,
-        name: str,
-        set_code: str,
-        collector_number: str,
-        image_size: str,
-        file_path: Path,
-        scryfall_uri: str | None = None,
-        artist: str | None = None,
-        face_index: int = 0,
-    ) -> None:
-        file_path_str = str(Path(file_path).resolve())
-
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            conn.execute(
-                """
-                INSERT OR REPLACE INTO card_images
-                (uuid, face_index, name, set_code, collector_number, image_size, file_path,
-                 downloaded_at, scryfall_uri, artist)
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-                (
-                    uuid,
-                    face_index,
-                    name,
-                    set_code,
-                    collector_number,
-                    image_size,
-                    file_path_str,
-                    datetime.now(UTC).isoformat(),
-                    scryfall_uri,
-                    artist,
-                ),
-            )
-            conn.commit()
-
-        key = (name.lower(), image_size)
-        with self._path_cache_lock:
-            self._path_cache.pop(key, None)
-
-    def get_cache_stats(self) -> dict[str, Any]:
-        with sqlite3.connect(self.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            total = conn.execute("SELECT COUNT(DISTINCT uuid) FROM card_images").fetchone()[0]
-            by_size = {}
-            for size in IMAGE_SIZES.values():
-                count = conn.execute(
-                    "SELECT COUNT(*) FROM card_images WHERE image_size = ?", (size,)
-                ).fetchone()[0]
-                by_size[size] = count
-
-            bulk_meta = conn.execute(
-                "SELECT downloaded_at, total_cards FROM bulk_data_meta WHERE id = 1"
-            ).fetchone()
-
-        return {
-            "unique_cards": total,
-            "by_size": by_size,
-            "bulk_data_date": bulk_meta[0] if bulk_meta else None,
-            "bulk_total_cards": bulk_meta[1] if bulk_meta else None,
-        }
-
-    def is_cached(self, uuid: str, size: str = "normal", face_index: int | None = 0) -> bool:
-        return self.get_image_by_uuid(uuid, size, face_index=face_index) is not None
-
-
-class BulkImageDownloader:
-    """High-throughput bulk image downloader using Scryfall data."""
-
-    def __init__(self, cache: CardImageCache, max_workers: int = SCRYFALL_MAX_DOWNLOAD_WORKERS):
-        self.cache = cache
-        self.max_workers = max_workers
-        self.session = requests.Session()
-        self.session.headers.update({"User-Agent": "MTGOMetagameCrawler/1.0"})
-
-    def _fetch_bulk_metadata(self) -> dict[str, Any]:
-        logger.info("Fetching bulk data metadata from Scryfall...")
-        resp = self.session.get(BULK_DATA_URL, timeout=SCRYFALL_REQUEST_TIMEOUT_SECONDS)
-        resp.raise_for_status()
-        return resp.json()
-
-    def fetch_card_by_name(self, name: str, set_code: str | None = None) -> dict[str, Any]:
-        params = {"exact": name}
-        if set_code:
-            params["set"] = set_code.lower()
-        resp = self.session.get(
-            SCRYFALL_CARD_NAMED_URL, params=params, timeout=SCRYFALL_REQUEST_TIMEOUT_SECONDS
-        )
-        resp.raise_for_status()
-        return resp.json()
-
-    def fetch_printings_by_name(self, name: str) -> list[dict[str, Any]]:
-        params = {"q": f'!"{name}"', "unique": "prints", "order": "released"}
-        results: list[dict[str, Any]] = []
-        url: str | None = SCRYFALL_CARD_SEARCH_URL
-        while url:
-            resp = self.session.get(url, params=params, timeout=SCRYFALL_REQUEST_TIMEOUT_SECONDS)
-            resp.raise_for_status()
-            payload = resp.json()
-            results.extend(payload.get("data", []))
-            url = payload.get("next_page")
-            params = None
-        return results
-
-    def download_card_image_by_name(
-        self, name: str, size: str = "normal", set_code: str | None = None
-    ) -> tuple[bool, str]:
-        card = self.fetch_card_by_name(name, set_code=set_code)
-        return self._download_single_image(card, size)
-
-    def _get_cached_bulk_data_record(self) -> tuple[str | None, str | None]:
-        with sqlite3.connect(self.cache.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS) as conn:
-            row = conn.execute(
-                "SELECT downloaded_at, bulk_data_uri FROM bulk_data_meta WHERE id = 1"
-            ).fetchone()
-        if row:
-            return row[0], row[1]
-        return None, None
-
-    def is_bulk_data_outdated(
-        self, max_staleness_seconds: int | None = None
-    ) -> tuple[bool, dict[str, Any]]:
-        metadata = self._fetch_bulk_metadata()
-        download_uri = metadata.get("download_uri")
-        updated_at = metadata.get("updated_at")
-
-        if not BULK_DATA_CACHE.exists():
-            return True, metadata
-
-        cached_updated, cached_uri = self._get_cached_bulk_data_record()
-        if updated_at and download_uri and cached_updated and cached_uri:
-            if updated_at == cached_updated and download_uri == cached_uri:
-                return False, metadata
-            return True, metadata
-
-        # Fallback to age-based check when the vendor metadata lacks timestamps/URIs
-        threshold = max_staleness_seconds or BULK_DATA_CACHE_FRESHNESS_SECONDS
-        try:
-            age_seconds = datetime.now().timestamp() - BULK_DATA_CACHE.stat().st_mtime
-            if age_seconds < threshold:
-                return False, metadata
-        except OSError:
-            pass
-
-        return True, metadata
-
-    def download_bulk_metadata(self, force: bool = False) -> tuple[bool, str]:
-        try:
-            metadata = self._fetch_bulk_metadata()
-        except Exception as exc:
-            logger.exception("Failed to fetch bulk data metadata")
-            return False, f"Error: {exc}"
-
-        download_uri = metadata.get("download_uri")
-        if not download_uri:
-            return False, "No download URI in bulk data response"
-
-        remote_updated_at = metadata.get("updated_at")
-        cached_updated, cached_uri = self._get_cached_bulk_data_record()
-        cache_matches = (
-            not force
-            and BULK_DATA_CACHE.exists()
-            and remote_updated_at
-            and cached_updated == remote_updated_at
-            and cached_uri == download_uri
-        )
-        if cache_matches:
-            logger.info("Using cached bulk data (vendor metadata is current)")
-            return True, "Using cached bulk data"
-
-        try:
-            logger.info(f"Downloading bulk data from {download_uri}")
-            logger.info(f"Size: {metadata.get('size', 0) / BYTES_PER_MB:.1f} MB")
-
-            # Download with progress
-            resp = self.session.get(
-                download_uri, stream=True, timeout=SCRYFALL_BULK_STREAM_TIMEOUT_SECONDS
-            )
-            resp.raise_for_status()
-
-            atomic_write_stream(
-                BULK_DATA_CACHE, resp.iter_content(chunk_size=SCRYFALL_DOWNLOAD_CHUNK_SIZE)
-            )
-
-            # Update database metadata (defer card count to avoid parsing 500MB file)
-            with sqlite3.connect(
-                self.cache.db_path, timeout=SQLITE_CONNECTION_TIMEOUT_SECONDS
-            ) as conn:
-                conn.execute(
-                    """
-                    INSERT OR REPLACE INTO bulk_data_meta (id, downloaded_at, total_cards, bulk_data_uri)
-                    VALUES (1, ?, ?, ?)
-                """,
-                    (
-                        remote_updated_at or datetime.now(UTC).isoformat(),
-                        0,
-                        download_uri,
-                    ),
-                )
-                conn.commit()
-
-            logger.info("Bulk data downloaded successfully")
-            return True, "Bulk data downloaded"
-
-        except Exception as exc:
-            logger.exception("Failed to download bulk data")
-            return False, f"Error: {exc}"
-
-    def _download_single_image(
-        self, card: dict[str, Any], size: str = "normal"
-    ) -> tuple[bool, str]:
-        uuid = card.get("id")
-        name = card.get("name", "Unknown")
-
-        if not uuid:
-            return False, f"No UUID for {name}"
-
-        card_faces = card.get("card_faces") or []
-        if card_faces:
-            return self._download_multi_face_card(card, card_faces, size)
-
-        success, message, _ = self._download_face_asset(
-            uuid=uuid,
-            face_index=0,
-            name=name,
-            image_uris=card.get("image_uris") or {},
-            size=size,
-            card=card,
-        )
-        return success, message
-
-    def _download_multi_face_card(
-        self, card: dict[str, Any], faces: list[dict[str, Any]], size: str
-    ) -> tuple[bool, str]:
-        uuid = card.get("id")
-        if not uuid:
-            return False, "Missing UUID for multi-face card"
-
-        downloaded = 0
-        front_path: Path | None = None
-        for idx, face in enumerate(faces):
-            face_name = face.get("name") or card.get("name", "Unknown")
-            image_uris = face.get("image_uris") or {}
-            success, _, file_path = self._download_face_asset(
-                uuid=uuid,
-                face_index=idx,
-                name=face_name,
-                image_uris=image_uris,
-                size=size,
-                card=card,
-            )
-            if success:
-                downloaded += 1
-                if idx == 0:
-                    front_path = file_path
-
-        # Store combined display name pointing to the front face
-        combined_name = card.get("name")
-        if combined_name and front_path:
-            self.cache.add_image(
-                uuid=uuid,
-                name=combined_name,
-                set_code=card.get("set", ""),
-                collector_number=card.get("collector_number", ""),
-                image_size=size,
-                file_path=front_path,
-                scryfall_uri=card.get("scryfall_uri"),
-                artist=card.get("artist"),
-                face_index=-1,
-            )
-
-        if downloaded == 0:
-            return False, f"No downloadable faces for {card.get('name', 'Unknown')}"
-        return True, f"Downloaded {downloaded} faces for {card.get('name', 'Unknown')}"
-
-    def _download_face_asset(
-        self,
-        uuid: str,
-        face_index: int,
-        name: str,
-        image_uris: dict[str, Any],
-        size: str,
-        card: dict[str, Any],
-    ) -> tuple[bool, str, Path | None]:
-        if self.cache.is_cached(uuid, size, face_index=face_index):
-            path = self.cache.get_image_by_uuid(uuid, size, face_index=face_index)
-            return True, f"Already cached: {name}", path
-
-        image_url = image_uris.get(size) or image_uris.get("normal")
-        if not image_url:
-            return False, f"No {size} image for {name}", None
-
-        try:
-            resp = self.session.get(image_url, timeout=SCRYFALL_REQUEST_TIMEOUT_SECONDS)
-            resp.raise_for_status()
-        except Exception as exc:
-            logger.debug(f"Failed to download {name}: {exc}")
-            return False, f"Error: {name} - {exc}", None
-
-        ext = "png" if size == "png" else "jpg"
-        filename = self._build_face_filename(uuid, face_index, ext)
-        file_path = self.cache.cache_dir / size / filename
-
-        try:
-            atomic_write_bytes(file_path, resp.content)
-        except Exception as exc:
-            logger.debug(f"Failed to write image {name}: {exc}")
-            return False, f"Error saving image for {name}: {exc}", None
-
-        self.cache.add_image(
-            uuid=uuid,
-            name=name,
-            set_code=card.get("set", ""),
-            collector_number=card.get("collector_number", ""),
-            image_size=size,
-            file_path=file_path,
-            scryfall_uri=card.get("scryfall_uri"),
-            artist=card.get("artist"),
-            face_index=face_index,
-        )
-        return True, f"Downloaded: {name}", file_path
-
-    @staticmethod
-    def _build_face_filename(uuid: str, face_index: int, ext: str) -> str:
-        if face_index <= 0:
-            return f"{uuid}.{ext}"
-        return f"{uuid}-f{face_index}.{ext}"
-
-    def download_all_images(
-        self,
-        size: str = "normal",
-        max_cards: int | None = None,
-        progress_callback: callable | None = None,
-    ) -> dict[str, Any]:
-        if not BULK_DATA_CACHE.exists():
-            return {
-                "success": False,
-                "error": "Bulk data not downloaded. Call download_bulk_metadata() first.",
-            }
-
-        try:
-            from utils.json_io import fast_load
-
-            with locked_path(BULK_DATA_CACHE):
-                cards_data = fast_load(BULK_DATA_CACHE)
-            if max_cards:
-                cards_data = cards_data[:max_cards]
-
-            total = len(cards_data)
-            completed = 0
-            successful = 0
-            failed = 0
-            skipped = 0
-
-            logger.info(f"Starting bulk download of {total} cards ({size} size)")
-
-            with ThreadPoolExecutor(max_workers=self.max_workers) as executor:
-                futures = {
-                    executor.submit(self._download_single_image, card, size): card
-                    for card in cards_data
-                }
-
-                for future in as_completed(futures):
-                    completed += 1
-                    try:
-                        success, message = future.result()
-                        if success:
-                            if "Already cached" in message:
-                                skipped += 1
-                            else:
-                                successful += 1
-                        else:
-                            failed += 1
-                            logger.debug(message)
-                    except Exception as exc:
-                        failed += 1
-                        logger.debug(f"Exception in download: {exc}")
-
-                    # Progress callback
-                    if progress_callback and completed % SCRYFALL_DOWNLOAD_PROGRESS_INTERVAL == 0:
-                        progress_callback(
-                            completed,
-                            total,
-                            f"{successful} downloaded, {skipped} cached, {failed} failed",
-                        )
-
-            logger.info(
-                f"Bulk download complete: {successful} downloaded, {skipped} cached, {failed} failed"
-            )
-
-            return {
-                "success": True,
-                "total": total,
-                "downloaded": successful,
-                "skipped": skipped,
-                "failed": failed,
-            }
-
-        except Exception as exc:
-            logger.exception("Bulk download failed")
-            return {"success": False, "error": str(exc)}
-
-
-def _collect_face_aliases(card: dict[str, Any], display_name: str) -> set[str]:
-    """Return alternate face names for MDFCs, split, and adventure cards."""
-    aliases: set[str] = set()
-    for raw_face in card.get("card_faces") or []:
-        face_name = (raw_face.get("name") or "").strip()
-        if face_name:
-            aliases.add(face_name)
-
-    if "//" in display_name:
-        for piece in display_name.split("//"):
-            face_name = piece.strip()
-            if face_name:
-                aliases.add(face_name)
-
-    display_key = display_name.strip().lower()
-    return {alias for alias in aliases if alias.lower() != display_key}
-
-
-def build_printing_index(
-    cards: list[dict[str, Any]],
-) -> tuple[dict[str, list[dict[str, Any]]], dict[str, int]]:
-    """Build a compact printing index from bulk card data."""
-    by_name: dict[str, list[dict[str, Any]]] = {}
-    total_printings = 0
-    for card in cards:
-        name = (card.get("name") or "").strip()
-        uuid = card.get("id")
-        if not name or not uuid:
-            continue
-        key = name.lower()
-        entry = {
-            "id": uuid,
-            "set": (card.get("set") or "").upper(),
-            "set_name": card.get("set_name") or "",
-            "collector_number": card.get("collector_number") or "",
-            "released_at": card.get("released_at") or "",
-            "flavor_text": card.get("flavor_text") or "",
-            "artist": card.get("artist") or "",
-        }
-        by_name.setdefault(key, []).append(entry)
-        for alias in _collect_face_aliases(card, name):
-            alias_key = alias.lower()
-            if alias_key == key:
-                continue
-            by_name.setdefault(alias_key, []).append(entry)
-        total_printings += 1
-
-    for entries in by_name.values():
-        entries.sort(key=lambda c: c.get("released_at") or "", reverse=True)
-
-    stats = {
-        "unique_names": len(by_name),
-        "total_printings": total_printings,
-    }
-    return by_name, stats
-
-
-def _load_printing_index_payload() -> dict[str, Any] | None:
-    """Load the cached card printings index if available."""
-    if not PRINTING_INDEX_CACHE.exists():
-        return None
-    try:
-        with locked_path(PRINTING_INDEX_CACHE):
-            raw = PRINTING_INDEX_CACHE.read_bytes()
-        payload = _printing_index_decoder.decode(raw)
-    except (msgspec.DecodeError, OSError) as exc:
-        logger.warning(f"Failed to read printings index cache: {exc}")
-        return None
-    if payload.version != PRINTING_INDEX_VERSION:
-        logger.info("Discarding printings index cache due to version mismatch")
-        return None
-    # Convert to plain dict so callers can use .get() / [] as before.
-    return msgspec.to_builtins(payload)
-
-
-def load_printing_index_payload() -> dict[str, Any] | None:
-    """Load the cached card printings index without rebuilding it."""
-    return _load_printing_index_payload()
-
-
-def ensure_printing_index_cache(force: bool = False) -> dict[str, Any]:
-    """Ensure a compact card printings index exists for fast wx lookups."""
-    IMAGE_CACHE_DIR.mkdir(parents=True, exist_ok=True)
-    existing = None if force else _load_printing_index_payload()
-    bulk_mtime = BULK_DATA_CACHE.stat().st_mtime if BULK_DATA_CACHE.exists() else None
-
-    if existing and (bulk_mtime is None or existing.get("bulk_mtime", 0) >= bulk_mtime):
-        return existing
-
-    if bulk_mtime is None:
-        raise FileNotFoundError("Bulk data cache not found; cannot build printings index")
-
-    logger.info("Building card printings index from bulk data…")
-    with locked_path(BULK_DATA_CACHE):
-        raw = BULK_DATA_CACHE.read_bytes()
-    cards = _bulk_cards_decoder.decode(raw)
-
-    by_name, stats = build_printing_index(cards)
-
-    payload = {
-        "version": PRINTING_INDEX_VERSION,
-        "generated_at": datetime.now(UTC).isoformat(),
-        "bulk_mtime": bulk_mtime,
-        "unique_names": stats["unique_names"],
-        "total_printings": stats["total_printings"],
-        "data": by_name,
-    }
-
-    try:
-        atomic_write_json(PRINTING_INDEX_CACHE, payload, separators=(",", ":"))
-        logger.info(
-            "Cached card printings index ({unique_names} names, {total_printings} printings)",
-            unique_names=payload["unique_names"],
-            total_printings=payload["total_printings"],
-        )
-    except Exception as exc:
-        logger.warning(f"Failed to write printings index cache: {exc}")
-
-    return payload
-
-
-# Singleton instance
-_cache_instance: CardImageCache | None = None
-
-
-def get_cache() -> CardImageCache:
-    """Get singleton CardImageCache instance."""
-    global _cache_instance
-    if _cache_instance is None:
-        _cache_instance = CardImageCache()
-    return _cache_instance
-
-
-def get_card_image(card_name: str, size: str = "normal") -> Path | None:
-    """Get card image from cache."""
-    cache = get_cache()
-    return cache.get_image_path(card_name, size)
-
-
-def download_bulk_images(
-    size: str = "normal", max_cards: int | None = None, progress_callback: callable | None = None
-) -> dict[str, Any]:
-    """High-level function to download all card images."""
-    cache = get_cache()
-    downloader = BulkImageDownloader(cache)
-
-    # Download metadata first
-    success, msg = downloader.download_bulk_metadata()
-    if not success:
-        return {"success": False, "error": msg}
-
-    # Download images
-    return downloader.download_all_images(size, max_cards, progress_callback)
-
-
-def get_cache_stats() -> dict[str, Any]:
-    """Get cache statistics."""
-    cache = get_cache()
-    return cache.get_cache_stats()
-
-
-__all__ = [
-    "CardImageCache",
-    "CardImageRequest",
-    "BulkImageDownloader",
-    "build_printing_index",
-    "PRINTING_INDEX_CACHE",
-    "get_cache",
-    "get_card_image",
-    "download_bulk_images",
-    "get_cache_stats",
-    "ensure_printing_index_cache",
-    "load_printing_index_payload",
-]
